Mag page (#32)
* Mag page in progress

* Sort data

* Added user detail on mag page

* apply review proposals + code cleanup

* menu filter by role

* * Added Mag page into mobile version
* Check the number of people with no experience for PIAF register
* Send mail to for long absences

* fixes

* Send email absence on form submit

* Validate PIAF

* Filter PIAF by validated

* dialog callback

* env var fix


Add info PIAF and hide reservation auto menu

Add PIAF info

New component PiafRow
