+++ src/pages/authenticated/Planning.tsx
@@ -7,8 +7,8 @@ import { isSameMonth } from "date-fns"
 import { Creneau } from "src/types/model"
 import { useDatePlanning } from "src/providers/datePlanning"
 import { PLANNING } from "src/graphql/queries"
-import Calendar from "src/components/Calendar"
 import Loader from "src/components/Loader"
+import Calendar from "src/components/Calendar"
 import IconsCaption from "src/components/IconsCaption"
 import { formatMonthYear } from "src/helpers/date"
 import { ErrorBlock } from "src/helpers/errors"
